requirements.txt
absl-py==2.3.1
anyio==4.9.0
apturl==0.5.2
async-timeout==4.0.3
attrs==25.3.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
bleak==0.22.3
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.3.2
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-fast==2.44.1
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.1.0
duplicity==0.8.21
exceptiongroup==1.2.2
fasteners==0.14.1
fierce==1.5.0
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.1
fsspec==2025.7.0
future==0.18.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.8
httplib2==0.20.2
httpx==0.26.0
idna==3.3
importlib-metadata==4.6.4
jax==0.6.2
jaxlib==0.6.2
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.5.1
kazam==1.4.5
keyring==23.5.0
kiwisolver==1.4.9
knockpy==5.3.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lightgbm==4.6.0
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.10.5
mediapipe==0.10.21
ml_dtypes==0.5.3
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
netaddr==0.8.0
netifaces==0.11.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.2.0
olefile==0.46
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pluginbase==1.0.1
protobuf==4.25.8
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-telegram-bot==20.8
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==0.9.1
scikit-learn==1.7.1
scipy==1.15.3
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
sentencepiece==0.2.1
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.3.1
splinter==0.21.0
sympy==1.14.0
systemd-python==234
threadpoolctl==3.6.0
tk==0.1.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
triton==3.4.0
typing_extensions==4.13.2
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.5.0
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
wafw00f==2.1.0
webencodings==0.5.1
xdg==5
xdot==1.2
xgboost==3.0.4
xkit==0.0.0
zipp==1.0.0
